requirements.txt
Flask==2.1.0
Pillow==9.1.0
opencv-python-headless==4.5.5.64
numpy==1.22.3